move reactWallComponent to reactWall/components
