--- app/components/dummy.js
@@ -1,1 +1,0 @@
-dummy.js
