+++ app/routes/reactWall/reactWallScene.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { View, StyleSheet } from 'react-native';
 import Header from '../../components/header';
-import ReactWall from '../../components/reactWallComponent';
+import ReactWall from './components/reactWallComponent';
 import PropTypes from 'prop-types';
 
 class ReactWallSceen extends Component {
